Move exceptions inside their relevant classes.
